stacktrail_backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9
django-cors-headers>=4.0
dnspython>=2.4.0
pyOpenSSL>=23.0.0
requests>=2.28.0
openai>=1.0.0
python-dotenv>=1.0.0
